BlockNumber, Nonce -> u32 (#1284)
* BlockNumber -> u32

* Interface types extend, not re-use

* Merge master, rebuild

* Merge typo

* Lazy create struct types (extended definitions)

* Cleanup tests for Nonce u32

* Remove polkadot u64 BlockNumber (substrate default)

* Warn on type dup & still clear

* Allow type overrides to work at any point

+++ packages/types/src/interfaces/runtime/definitions.ts
@@ -12,7 +12,7 @@ export default {
     Balance: 'u128',
     BalanceOf: 'Balance',
     Block: 'GenericBlock',
-    BlockNumber: 'u64',
+    BlockNumber: 'u32',
     Call: 'GenericCall',
     ConsensusEngineId: 'GenericConsensusEngineId',
     Digest: 'GenericDigest',
@@ -28,7 +28,7 @@ export default {
       extrinsicsRoot: 'Hash',
       digest: 'Digest'
     },
-    Index: 'u64',
+    Index: 'u32',
     Justification: 'Bytes',
     KeyValue: '(StorageKey, StorageData)',
     KeyTypeId: 'u32',
